+++ MHD/C_src/KEMO_GL/set_PSF_patches_to_buf.c
@@ -98,10 +98,10 @@ void set_psf_map_to_buf(int ist_psf, int ied_psf, struct psf_data **psf_s,
 }
 
 
-int count_psf_arrows_to_buf(int ncorner, struct psf_data *psf_s, struct psf_menu_val *psf_m){
+long count_psf_arrows_to_buf(int ncorner, struct psf_data *psf_s, struct psf_menu_val *psf_m){
 	int inod;
 	
-	int inum_buf = 0;
+	long inum_buf = 0;
 	for (inod = 0; inod < psf_s->nnod_viz; inod++) {
 		if (inod % psf_m->increment_vect == 0) {
 			if(psf_s->norm_nod[inod][0] != 0.0
