+++ MHD/C_src/Metal/KemoViewMetalBuffers.h
@@ -22,27 +22,27 @@
 
 @interface KemoViewMetalBuffers : NSObject
 
-- (int) setMetalVertexs:(id<MTLDevice> _Nonnull *_Nonnull) device
-                 buffer:(struct gl_strided_buffer * _Nonnull) buf
-                 vertex:(id<MTLBuffer> _Nonnull *_Nonnull)  vertices;
+- (NSUInteger) setMetalVertexs:(id<MTLDevice> _Nonnull *_Nonnull) device
+                        buffer:(struct gl_strided_buffer * _Nonnull) buf
+                        vertex:(id<MTLBuffer> _Nonnull *_Nonnull)  vertices;
 
-- (int) setPSFTexture:(id<MTLDevice> _Nonnull *_Nonnull) device
-               buffer:(struct gl_strided_buffer *_Nonnull) buf
-                image:(struct kemo_PSF_texure *_Nonnull) psf_texure
-               vertex:(id<MTLBuffer> _Nonnull *_Nonnull)  vertices
-               texure:(id<MTLTexture> _Nonnull *_Nonnull) texture;
+- (NSUInteger) setPSFTexture:(id<MTLDevice> _Nonnull *_Nonnull) device
+                      buffer:(struct gl_strided_buffer *_Nonnull) buf
+                       image:(struct kemo_PSF_texure *_Nonnull) psf_texure
+                      vertex:(id<MTLBuffer> _Nonnull *_Nonnull)  vertices
+                      texure:(id<MTLTexture> _Nonnull *_Nonnull) texture;
 
-- (int) setTextBoxTexture:(id<MTLDevice> _Nonnull *_Nonnull) device
-                   buffer:(struct gl_strided_buffer *_Nonnull) buf
-                    image:(struct line_text_image *_Nonnull) img
-                   vertex:(id<MTLBuffer> _Nonnull *_Nonnull)  vertices
-                   texure:(id<MTLTexture> _Nonnull *_Nonnull) texture;
+- (NSUInteger) setTextBoxTexture:(id<MTLDevice> _Nonnull *_Nonnull) device
+                          buffer:(struct gl_strided_buffer *_Nonnull) buf
+                           image:(struct line_text_image *_Nonnull) img
+                          vertex:(id<MTLBuffer> _Nonnull *_Nonnull)  vertices
+                          texure:(id<MTLTexture> _Nonnull *_Nonnull) texture;
 
-- (int) setCubeVertexs:(id<MTLDevice> _Nonnull *_Nonnull) device
-                buffer:(struct gl_strided_buffer *_Nonnull) buf
-              indexbuf:(struct gl_index_buffer *_Nonnull) index_buf
-                vertex:(id<MTLBuffer> _Nonnull *_Nonnull) vertices
-                 index:(id<MTLBuffer> _Nonnull *_Nonnull) indices;
+- (NSUInteger) setCubeVertexs:(id<MTLDevice> _Nonnull *_Nonnull) device
+                       buffer:(struct gl_strided_buffer *_Nonnull) buf
+                     indexbuf:(struct gl_index_buffer *_Nonnull) index_buf
+                       vertex:(id<MTLBuffer> _Nonnull *_Nonnull) vertices
+                        index:(id<MTLBuffer> _Nonnull *_Nonnull) indices;
 
 - (void)setAnaglyphTexture:(id<MTLDevice> _Nonnull *_Nonnull) device
                     buffer:(struct gl_strided_buffer *_Nonnull) buf
